requirements.txt
mediapipe>=0.10.0
opencv-python>=4.8.0
numpy>=1.24.0
ultralytics>=8.0.0
pytest>=7.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
websockets>=12.0
twilio>=8.0.0
python-dotenv>=1.0.0
